+++ tests/test_slack.py
@@ -17,8 +17,11 @@ def send_message(config, client, message):
 # The actual test
 @pytest.mark.integration
 def test_slack_integration():
-    with open("files/config.yml") as f:
+    with open("files/config_template.yml") as f:
         config = yaml.safe_load(f)
+
+    if "your-slack-bot-token" in config.get("SLACK")["bot_token"]:
+        pytest.skip("Slack config is not set up for integration test.")
 
     client = WebClient(token=config.get("SLACK")["bot_token"])